backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx

torch
torchvision
Pillow

langchain
langchain-openai
langchain-community
faiss-cpu
tiktoken
python-dotenv
langchain-text-splitters

markdown
unstructured
python-dotenv

ultralytics
